requirements.txt
-i https://pypi.org/simple
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
click==8.1.7
ete3==3.1.3
idna==3.4 ; python_version >= '3.5'
numpy==1.25.2
pyyaml==6.0.1 ; python_version >= '3.6'
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7 ; python_version < '3.12' and platform_python_implementation == 'CPython'
urllib3==2.0.4 ; python_version >= '3.7'
yamale==4.0.4
